Refactor project 8

+++ SwapyAPI/Swapy.BLL/Domain/Auth/CommandHandlers/LoginCommandHandler.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Swapy.BLL.Domain.Auth.Commands;
 using Swapy.BLL.Interfaces;
-using Swapy.Common.DTO;
+using Swapy.Common.DTO.Auth.Responses;
 using Swapy.Common.Entities;
 using Swapy.Common.Exceptions;
 using Swapy.DAL.Interfaces;
@@ -13,16 +13,16 @@ namespace Swapy.BLL.Domain.Auth.CommandHandlers
 {
     public class LoginCommandHandler : IRequestHandler<LoginCommand, AuthResponseDTO>
     {
-        private readonly IUserTokenService _userTokenService;
         private readonly UserManager<User> _userManager;
         private readonly SignInManager<User> _signInManager;
+        private readonly IUserTokenService _userTokenService;
         private readonly IUserTokenRepository _userTokenRepository;
 
         public LoginCommandHandler(UserManager<User> userManager, SignInManager<User> signInManager, IUserTokenRepository userTokenRepository, IUserTokenService userTokenService)
         {
             _userManager = userManager;
-            _userTokenService = userTokenService;
             _signInManager = signInManager;
+            _userTokenService = userTokenService;
             _userTokenRepository = userTokenRepository;
         }
 
@@ -39,7 +39,7 @@ namespace Swapy.BLL.Domain.Auth.CommandHandlers
             var result = await _signInManager.CheckPasswordSignInAsync(user, request.Password, false);
             if (!result.Succeeded) throw new AuthenticationException("Invalid email or password");
 
-            await _userTokenRepository.DeleteByIdAsync(user.UserTokenId);
+            if (!string.IsNullOrEmpty(user.UserTokenId)) await _userTokenRepository.DeleteByIdAsync(user.UserTokenId);
 
             var accessToken = await _userTokenService.GenerateJwtToken(user.Id, user.Email, user.FirstName, user.LastName);
             var refreshToken = await _userTokenService.GenerateRefreshToken();
